euler: move Limiter and WaveSpeedEstimator definitions into their headers
