--- source/euler/limiter.template.h
@@ -1,328 +1,0 @@
-//
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-// Copyright (C) 2020 - 2025 by the ryujin authors
-//
-
-#pragma once
-
-#include "limiter.h"
-// #define DEBUG_OUTPUT_LIMITER
-
-namespace ryujin
-{
-  namespace Euler
-  {
-    template <int dim, typename Number>
-    std::tuple<Number, bool>
-    LimiterView<dim, Number>::limit(const Bounds &bounds,
-                                    const state_type &U,
-                                    const state_type &P,
-                                    const Number t_min /* = Number(0.) */,
-                                    const Number t_max /* = Number(1.) */) const
-    {
-      bool success = true;
-      Number t_r = t_max;
-
-      constexpr ScalarNumber eps = std::numeric_limits<ScalarNumber>::epsilon();
-      const auto small = view_.vacuum_state_relaxation_small();
-      const auto large = view_.vacuum_state_relaxation_large();
-      const ScalarNumber relax_small = ScalarNumber(1. + small * eps);
-      const ScalarNumber relax = ScalarNumber(1. + large * eps);
-
-      /*
-       * First limit the density rho.
-       *
-       * See [Guermond, Nazarov, Popov, Thomas] (4.8):
-       */
-
-      {
-        const auto &rho_U = view_.density(U);
-        const auto &rho_P = view_.density(P);
-
-        const auto &rho_min = std::get<0>(bounds);
-        const auto &rho_max = std::get<1>(bounds);
-
-        /*
-         * Verify that rho_U is within bounds. This property might be
-         * violated for relative CFL numbers larger than 1.
-         */
-        const auto test_min = view_.filter_vacuum_density(
-            std::max(Number(0.), rho_U - relax * rho_max));
-        const auto test_max = view_.filter_vacuum_density(
-            std::max(Number(0.), rho_min - relax * rho_U));
-        if (!(test_min == Number(0.) && test_max == Number(0.))) {
-#ifdef DEBUG_OUTPUT
-          std::cout << std::fixed << std::setprecision(16);
-          std::cout << "Bounds violation: low-order density (critical)!"
-                    << "\n\t\trho min:         " << rho_min
-                    << "\n\t\trho min (delta): "
-                    << negative_part(rho_U - rho_min)
-                    << "\n\t\trho:             " << rho_U
-                    << "\n\t\trho max (delta): "
-                    << positive_part(rho_U - rho_max)
-                    << "\n\t\trho max:         " << rho_max << "\n"
-                    << std::endl;
-#endif
-          success = false;
-        }
-
-        const Number denominator =
-            ScalarNumber(1.) / (std::abs(rho_P) + eps * rho_max);
-
-        constexpr auto lt = dealii::SIMDComparison::less_than;
-
-        t_r = dealii::compare_and_apply_mask<lt>( //
-            rho_max,
-            rho_U + t_r * rho_P,
-            /*
-             * rho_P is positive.
-             *
-             * Note: Do not take an absolute value here. If we are out of
-             * bounds we have to ensure that t_r is set to t_min.
-             */
-            (rho_max - rho_U) * denominator,
-            t_r);
-
-        t_r = dealii::compare_and_apply_mask<lt>( //
-            rho_U + t_r * rho_P,
-            rho_min,
-            /*
-             * rho_P is negative.
-             *
-             * Note: Do not take an absolute value here. If we are out of
-             * bounds we have to ensure that t_r is set to t_min.
-             */
-            (rho_U - rho_min) * denominator,
-            t_r);
-
-        /*
-         * Ensure that t_min <= t <= t_max. This might not be the case if
-         * rho_U is outside the interval [rho_min, rho_max]. Furthermore,
-         * the quotient we take above is prone to numerical cancellation in
-         * particular in the second pass of the limiter when rho_P might be
-         * small.
-         */
-        t_r = std::min(t_r, t_max);
-        t_r = std::max(t_r, t_min);
-
-#ifdef DEBUG_EXPENSIVE_BOUNDS_CHECK
-        /*
-         * Verify that the new state is within bounds:
-         */
-        const auto rho_new = view_.density(U + t_r * P);
-        const auto test_new_min = view_.filter_vacuum_density(
-            std::max(Number(0.), rho_new - relax * rho_max));
-        const auto test_new_max = view_.filter_vacuum_density(
-            std::max(Number(0.), rho_min - relax * rho_new));
-        if (!(test_new_min == Number(0.) && test_new_max == Number(0.))) {
-#ifdef DEBUG_OUTPUT
-          std::cout << std::fixed << std::setprecision(16);
-          std::cout << "Bounds violation: high-order density!"
-                    << "\n\t\trho min:         " << rho_min
-                    << "\n\t\trho min (delta): "
-                    << negative_part(rho_new - rho_min)
-                    << "\n\t\trho:             " << rho_new
-                    << "\n\t\trho max (delta): "
-                    << positive_part(rho_new - rho_max)
-                    << "\n\t\trho max:         " << rho_max << "\n"
-                    << std::endl;
-#endif
-          success = false;
-        }
-#endif
-      }
-
-      /*
-       * Then limit the specific entropy:
-       *
-       * See [Guermond, Nazarov, Popov, Thomas], Section 4.6 + Section 5.1:
-       */
-
-      Number t_l = t_min; // good state
-
-      const ScalarNumber gamma = view_.gamma();
-      const ScalarNumber gp1 = gamma + ScalarNumber(1.);
-
-      {
-        /*
-         * Prepare a quadratic Newton method:
-         *
-         * Given initial limiter values t_l and t_r with psi(t_l) > 0 and
-         * psi(t_r) < 0 we try to find t^\ast with psi(t^\ast) \approx 0.
-         *
-         * Here, psi is a 3-convex function obtained by scaling the specific
-         * entropy s:
-         *
-         *   psi = \rho ^ {\gamma + 1} s
-         *
-         * (s in turn was defined as s =\varepsilon \rho ^{-\gamma}, where
-         * \varepsilon = (\rho e) is the internal energy.)
-         */
-
-        const auto &s_min = std::get<2>(bounds);
-
-#ifdef DEBUG_OUTPUT_LIMITER
-        std::cout << std::endl;
-        std::cout << std::fixed << std::setprecision(16);
-        std::cout << "t_l: (start) " << t_l << std::endl;
-        std::cout << "t_r: (start) " << t_r << std::endl;
-#endif
-
-        for (unsigned int n = 0; n < limiter_.newton_max_iterations(); ++n) {
-
-          const auto U_r = U + t_r * P;
-          const auto rho_r = view_.density(U_r);
-          const auto rho_r_gamma = ryujin::pow(rho_r, gamma);
-          const auto rho_e_r = view_.internal_energy(U_r);
-
-          auto psi_r =
-              relax_small * rho_r * rho_e_r - s_min * rho_r * rho_r_gamma;
-
-#ifndef DEBUG_EXPENSIVE_BOUNDS_CHECK
-          /*
-           * If psi_r > 0 the right state is fine, force returning t_r by
-           * setting t_l = t_r:
-           */
-          t_l = dealii::compare_and_apply_mask<
-              dealii::SIMDComparison::greater_than>(
-              psi_r, Number(0.), t_r, t_l);
-
-          /*
-           * If we have set t_l = t_r everywhere then all states state U_r
-           * with t_r obey the specific entropy inequality and we can
-           * break.
-           *
-           * This is a very important optimization: Only for 1 in (25 to
-           * 50) cases do we actually need to limit on the specific entropy
-           * because one of the right states failed. So we can skip
-           * constructing the left state U_l, which is expensive.
-           *
-           * This implies unfortunately that we might not accurately report
-           * whether the low_order update U itself obeyed bounds because
-           * U_r = U + t_r * P pushed us back into bounds. We thus skip
-           * this shortcut if `DEBUG_EXPENSIVE_BOUNDS_CHECK` is set.
-           */
-          if (t_l == t_r) {
-#ifdef DEBUG_OUTPUT_LIMITER
-            std::cout << "shortcut: t_l == t_r" << std::endl;
-            std::cout << "psi_l:       " << psi_l << std::endl;
-            std::cout << "psi_r:       " << psi_r << std::endl;
-            std::cout << "t_l: (  " << n << "  ) " << t_l << std::endl;
-            std::cout << "t_r: (  " << n << "  ) " << t_r << std::endl;
-#endif
-            break;
-          }
-#endif
-
-          const auto U_l = U + t_l * P;
-          const auto rho_l = view_.density(U_l);
-          const auto rho_l_gamma = ryujin::pow(rho_l, gamma);
-          const auto rho_e_l = view_.internal_energy(U_l);
-
-          auto psi_l =
-              relax_small * rho_l * rho_e_l - s_min * rho_l * rho_l_gamma;
-
-          /*
-           * Verify that the left state is within bounds. This property might
-           * be violated for relative CFL numbers larger than 1.
-           */
-          const auto lower_bound =
-              (ScalarNumber(1.) - relax) * s_min * rho_l * rho_l_gamma;
-          if (n == 0 &&
-              !(std::min(Number(0.), psi_l - lower_bound) == Number(0.))) {
-#ifdef DEBUG_OUTPUT
-            std::cout << std::fixed << std::setprecision(16);
-            std::cout
-                << "Bounds violation: low-order specific entropy (critical)!\n";
-            std::cout << "\t\tPsi left: 0 <= " << psi_l << "\n" << std::endl;
-#endif
-            success = false;
-          }
-
-#ifdef DEBUG_EXPENSIVE_BOUNDS_CHECK
-          /*
-           * If psi_r > 0 the right state is fine, force returning t_r by
-           * setting t_l = t_r:
-           */
-          t_l = dealii::compare_and_apply_mask<
-              dealii::SIMDComparison::greater_than>(
-              psi_r, Number(0.), t_r, t_l);
-#endif
-
-          /*
-           * Break if the window between t_l and t_r is within the prescribed
-           * tolerance:
-           */
-          const Number tolerance(limiter_.newton_tolerance());
-          if (std::max(Number(0.), t_r - t_l - tolerance) == Number(0.)) {
-#ifdef DEBUG_OUTPUT_LIMITER
-            std::cout << "break: t_l and t_r within tolerance" << std::endl;
-            std::cout << "psi_l:       " << psi_l << std::endl;
-            std::cout << "psi_r:       " << psi_r << std::endl;
-            std::cout << "t_l: (  " << n << "  ) " << t_l << std::endl;
-            std::cout << "t_r: (  " << n << "  ) " << t_r << std::endl;
-#endif
-            break;
-          }
-
-          /* We got unlucky and have to perform a Newton step: */
-
-          const auto drho = view_.density(P);
-          const auto drho_e_l = view_.internal_energy_derivative(U_l) * P;
-          const auto drho_e_r = view_.internal_energy_derivative(U_r) * P;
-          const auto dpsi_l =
-              rho_l * drho_e_l + (rho_e_l - gp1 * s_min * rho_l_gamma) * drho;
-          const auto dpsi_r =
-              rho_r * drho_e_r + (rho_e_r - gp1 * s_min * rho_r_gamma) * drho;
-
-          quadratic_newton_step(
-              t_l, t_r, psi_l, psi_r, dpsi_l, dpsi_r, Number(-1.));
-
-#ifdef DEBUG_OUTPUT_LIMITER
-          std::cout << "psi_l:       " << psi_l << std::endl;
-          std::cout << "psi_r:       " << psi_r << std::endl;
-          std::cout << "dpsi_l:      " << dpsi_l << std::endl;
-          std::cout << "dpsi_r:      " << dpsi_r << std::endl;
-          std::cout << "t_l: (  " << n << "  ) " << t_l << std::endl;
-          std::cout << "t_r: (  " << n << "  ) " << t_r << std::endl;
-#endif
-        }
-
-#ifdef DEBUG_EXPENSIVE_BOUNDS_CHECK
-        /*
-         * Verify that the new state is within bounds:
-         */
-        {
-          const auto U_new = U + t_l * P;
-          const auto rho_new = view_.density(U_new);
-          const auto rho_new_gamma = ryujin::pow(rho_new, gamma);
-          const auto rho_e_new = view_.internal_energy(U_new);
-
-          auto psi_new = relax_small * rho_new * rho_e_new -
-                         s_min * rho_new * rho_new_gamma;
-
-          const auto lower_bound =
-              (ScalarNumber(1.) - relax) * s_min * rho_new * rho_new_gamma;
-
-          const bool e_valid = std::min(Number(0.), rho_e_new) == Number(0.);
-          const bool psi_valid =
-              std::min(Number(0.), psi_new - lower_bound) == Number(0.);
-
-          if (!e_valid || !psi_valid) {
-#ifdef DEBUG_OUTPUT
-            std::cout << std::fixed << std::setprecision(16);
-            std::cout << "Bounds violation: high-order specific entropy!\n";
-            std::cout << "\t\trho e: 0 <= " << rho_e_new << "\n";
-            std::cout << "\t\tPsi:   0 <= " << psi_new << "\n" << std::endl;
-#endif
-            success = false;
-          }
-        }
-#endif
-      }
-
-      return {t_l, success};
-    }
-
-  } // namespace Euler
-} // namespace ryujin
