+++ tests/euler/wave_speed_estimator.cc
@@ -6,7 +6,6 @@
 #define DEBUG_WAVE_SPEED_ESTIMATOR
 #include <simd.h>
 #include <wave_speed_estimator.h>
-#include <wave_speed_estimator.template.h>
 
 #ifndef NEWTON_ITERATIONS
 #define NEWTON_ITERATIONS "0"
